chore: refactor interceptor
