+++ plugin/host/service.go
@@ -18,9 +18,6 @@ import (
 	"github.com/google/uuid"
 	lru "github.com/hashicorp/golang-lru/v2/expirable"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/metadata"
-	"google.golang.org/grpc/status"
 
 	v1 "github.com/flanksource/incident-commander/api/v1"
 	"github.com/flanksource/incident-commander/auth"
@@ -69,56 +66,6 @@ func New(ctx dutyContext.Context, pluginID uuid.UUID) *Service {
 // Register exposes the service on the given gRPC server.
 func (s *Service) Register(g *grpc.Server) {
 	pluginpb.RegisterHostServiceServer(g, s)
-}
-
-func (s *Service) UnaryServerInterceptor() grpc.UnaryServerInterceptor {
-	return func(ctx context.Context, req any, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (any, error) {
-		if !requiresInvocation(info.FullMethod) {
-			return handler(ctx, req)
-		}
-
-		invocationCtx, err := s.contextWithInvocation(ctx)
-		if err != nil {
-			return nil, err
-		}
-		return handler(invocationCtx, req)
-	}
-}
-
-func requiresInvocation(method string) bool {
-	switch method {
-	case pluginpb.HostService_GetConfigItem_FullMethodName,
-		pluginpb.HostService_ListConfigs_FullMethodName,
-		pluginpb.HostService_GetConnection_FullMethodName:
-		return true
-	default:
-		return false
-	}
-}
-
-func (s *Service) contextWithInvocation(ctx context.Context) (context.Context, error) {
-	md, ok := metadata.FromIncomingContext(ctx)
-	if !ok {
-		return nil, status.Error(codes.Unauthenticated, "plugin invocation token is required")
-	}
-
-	values := md.Get(pluginpb.PluginInvocationTokenMetadataKey)
-	if len(values) == 0 || values[0] == "" {
-		return nil, status.Error(codes.Unauthenticated, "plugin invocation token is required")
-	}
-
-	claims, err := auth.VerifyPluginInvocationToken(values[0], s.pluginID)
-	if err != nil {
-		return nil, status.Errorf(codes.Unauthenticated, "invalid plugin invocation token: %v", err)
-	}
-
-	baseCtx := s.ctx.Wrap(ctx)
-	var person models.Person
-	if err := baseCtx.DB().WithContext(ctx).Where("id = ?", claims.Subject).First(&person).Error; err != nil {
-		return nil, status.Errorf(codes.Unauthenticated, "plugin invocation subject %s: %v", claims.Subject, err)
-	}
-
-	return baseCtx.WithUser(&person), nil
 }
 
 func (s *Service) GetConfigItem(ctx context.Context, req *pluginpb.GetConfigItemRequest) (*pluginpb.ConfigItem, error) {
